Commit
231110_1427
[2021 UPDATE] USER PROFILE
END

+++ src/controllers/videoController.js
@@ -1,32 +1,44 @@
 import Video from "../models/Video.js";
+import User from "../models/User.js";
 
 
 export const home = async(req, res) => {
     try {
         const videos = await Video.find({}).sort({createdAt: "desc"});
-        console.log("⭐ (home)videos : ", videos);
+        // console.log("⭐ (home)videos : ", videos);
         return res.render("home", { pageTitle: "Home", videos: videos });
     } catch(error) {
         return res.status(400).render("❌ (home)Server-Error", { error });
     };
 };
 export const watch = async(req, res) => {
-    const id = req.params.id;
-    const video = await Video.findById(id);
+    const { id } = req.params;
+    const video = await  Video.findById(id).populate("owner");
+    // const owner = await User.findById(video.owner);
+    console.log("VIDEO : ", video);
     if(!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
-    return res.render("watch", { pageTitle: video.title, video: video });
+    return res.render("watch", { pageTitle: video.title, video: video,  });  // owner
 };
 export const getEdit = async(req, res) => {
-    const id = req.params.id;
+    const { id } = req.params;
+    const { user: {_id} } = req.session;
     const video = await Video.findById(id)
     if(!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." })
     }
+    console.log("VIDEO.OWNER : ", video.owner),
+    console.log("_ID : ", _id);
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video: video });
 };
 export const postEdit = async(req, res) => {
+    const {
+        user: { _id },
+    } = req.session;
     const id = req.params.id;
     const title = req.body.title;
     const description = req.body.title;
@@ -34,6 +46,9 @@ export const postEdit = async(req, res) => {
     const video = await Video.exists({ _id: id });
     if(!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title: title,
@@ -46,19 +61,33 @@ export const getUpload = (req, res) => {
     return res.render("Upload", { pageTitle: "Upload Video" });
 };
 export const postUpload = async(req, res) => {
-    const file = req.file;
-    const title = req.body.title;
-    const description = req.body.description;
-    const hashtags = req.body.hashtags;
+    const {
+        session: {
+            user: _id
+        },
+        body: { title, description, hashtags },
+        file,
+    } = req;
     try {
-        const video = new Video({
-            title: title,
-            description: description,
+        const newVideo = await Video.create({
+            title,
+            description,
             fileUrl: file.path,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
-        await video.save();
-        console.log(`⭐ (postUpload)Saved Video : `, video);
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
+        // const video = new Video({
+        //     title: title,
+        //     description: description,
+        //     fileUrl: file.path,
+        //     owner: _id,
+        //     hashtags: Video.formatHashtags(hashtags),
+        // });
+        // await video.save();
+        // console.log(`⭐ (postUpload)Saved Video : `, video);
         return res.redirect("/");
     } catch(error) {
         console.log(`❌ (postUpload)Server-Error : \n${ error }`);
@@ -67,7 +96,20 @@ export const postUpload = async(req, res) => {
 };
 
 export const deleteVideo = async(req, res) => {
-    const id = req.params.id;
+    // const { id } = req.params;
+    const {
+        session: {
+            user: { _id }
+        },
+        params: {id}
+    } = req;
+    const video = await Video.findById(id);
+    if(!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     // delete video
     return res.redirect("/");
